+++ bitframe-client/viewmodels/src/commonMain/kotlin/bitframe/authentication/signin/SignInViewModel.kt
@@ -14,24 +14,27 @@ import kotlin.js.JsExport
 
 class SignInViewModel(
     private val service: SignInService
-) : ViewModel<SignInIntent, SignInState>(SignInState(SignInFormFields(), null)) {
+) : ViewModel<SignInIntent, SignInState>(SignInState.Form(SignInFormFields(), null)) {
 
     private val recoveryTime = 3000
 
-    val console = logger()
-
     override fun CoroutineScope.execute(i: SignInIntent): Any = when (i) {
         is SignInIntent.Submit -> signIn(i)
+        is SignInIntent.Resolve -> resolve(i)
+    }
+
+    private fun resolve(i: SignInIntent.Resolve) {
+        val state = ui.value as SignInState.Conundrum
+        service.resolve(i.space)
     }
 
     private fun CoroutineScope.signIn(i: SignInIntent.Submit) = launch {
-        val state = ui.value.copy(status = Loading("Signing you in, please wait . . ."))
+        val state = ui.value as SignInState.Form
         flow {
-            emit(state)
+            emit(state.copy(status = Loading("Signing you in, please wait . . .")))
             val conundrum = service.signIn(i.credentials).await()
             if (conundrum.spaces.size > 1) {
-                console.warn("User has more than one Space, ")
-                emit(state.copy(status = Success("Logged in successfully")))
+                emit(SignInState.Conundrum(conundrum.user, conundrum.spaces))
             } else {
                 emit(state.copy(status = Success("Logged in successfully")))
             }