+++ bitframe-client/viewmodels/src/jsMain/kotlin/bitframe/authentication/signin/exports/SignInScope.kt
@@ -4,6 +4,7 @@
 package bitframe.authentication.signin.exports
 
 import bitframe.authentication.signin.*
+import bitframe.authentication.spaces.Space
 import useEventHandler
 import viewmodel.ViewModel
 
@@ -13,6 +14,10 @@ class SignInScope(service: SignInService) {
 
     val submit = { cred: SignInCredentials ->
         viewModel.post(SignInIntent.Submit(cred.toSignInCredentials()))
+    }
+
+    val resolve = { space: Space ->
+        viewModel.post(SignInIntent.Resolve(space))
     }
 
     val useSignInEvent: (callback: (Session.SignedIn) -> Unit) -> Unit = {