+++ pi-monitor/pi-monitor-services/stub/src/commonMain/kotlin/pimonitor/PiMonitorServiceStub.kt
@@ -3,9 +3,11 @@ package pimonitor
 import bitframe.authentication.AuthenticationDaoProvider
 import bitframe.authentication.InMemoryAuthenticationDaoProvider
 import bitframe.authentication.signin.SignInServiceImpl
+import bitframe.authentication.spaces.SpacesServiceImpl
 import bitframe.authentication.users.UsersService
 import bitframe.authentication.users.UsersServiceImpl
 import bitframe.events.InMemoryEventBus
+import pimonitor.authentication.signup.SignUpParams
 import pimonitor.authentication.signup.SignUpServiceImpl
 import pimonitor.evaluation.businesses.BusinessServiceImpl
 import pimonitor.monitored.MonitoredBusinessDaoInMemory
@@ -23,7 +25,8 @@ fun PiMonitorServiceStub(
     val monitorDao = MonitorDaoInMemory(config = daoConfig)
     val monitoredBusinessDao = MonitoredBusinessDaoInMemory(config = daoConfig)
     val monitorsService = MonitorsServiceImpl(signInService.session, monitorDao, config)
-    return object : PiMonitorService(
+    val service = object : PiMonitorService(
+        spaces = SpacesServiceImpl(config, provider.spaces, provider.users),
         users = usersService,
         signIn = signInService,
         signUp = SignUpServiceImpl(monitorDao, usersService, bus, config),
@@ -31,4 +34,6 @@ fun PiMonitorServiceStub(
         businesses = BusinessServiceImpl(bus, monitoredBusinessDao, monitorsService, config),
         bus = bus
     ) {}
+    service.populateTestEntities()
+    return service
 }