+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+# 0.0.32
+
+## PiMonitor Authentication
+
+- [[PM164]](https://github.com/picortex/bitframe/issues/164) Fix SingUp Dropdown feature
+
+## Bitframe Authentication
+
+- [[BF163]](https://github.com/picortex/bitframe/issues/163) Optimize SignIn Flow
+- [[BF165]](https://github.com/picortex/bitframe/issues/165) Add Valid Users for Test & Development Purposes
+
 # 0.0.31
 
 ## PiMonitor Client SDK Core